Fixes and improvements to piece movement
Fixed piece movement so strings now translate into the correct
positions.
Added a check on addPiece to make sure we aren't trying to move on top
of one of our own units
Fixed bug that caused segfault when empty string entered as command.
Added check to movePiece using getMovable but seems to prevent all
movemenet so commented out for now.